+++ src/components/InkEditor/components/ToolbarButton/ToolbarButton.tsx
@@ -1,9 +1,14 @@
-import type { FC } from 'react';
+import type { FC, MouseEvent } from 'react';
 import type { ToolbarButtonProps } from '../../../../types';
 import { Button } from '@common-components';
 
 export const ToolbarButton: FC<ToolbarButtonProps> = (props) => {
   const { icon, title, active = false, onClick, disabled = false } = props;
+
+  const stopMouseDown = (event: MouseEvent<HTMLButtonElement>) => {
+    event.preventDefault();
+  };
+
   return (
     <Button
       title={title}
@@ -11,6 +16,7 @@ export const ToolbarButton: FC<ToolbarButtonProps> = (props) => {
       aria-pressed={active}
       disabled={disabled}
       active={active}
+      onMouseDown={stopMouseDown}
       onClick={onClick}
     >
       {icon}